separate converter from materializer
